requirements.txt
# core deeplearning
# --index-url https://download.pytorch.org/whl/cu117
# torch==2.0.0+cu117
# torchvision==0.15.1+cu117

# management
# ultralytics @ git+https://github.com/muhammadAgfian96/ultralytics.git@5560391dd94556f0f9aaefbdf1bf8b887b8bd0a2
# ultralytics==8.0.114
ultralytics @ git+https://github.com/muhammadAgfian96/ultralytics.git@feat/disable-#egg=ultralytics
clearml==1.11.0

# images
opencv-python>=4.1.2
Pillow
numpy
pandas

# data
albumentations 
label-studio-sdk==0.0.24
label-studio-converter==0.0.53
cvat-sdk==2.4.4
roboflow
boto3
gsutil 
requests
pycocotools>=2.0.6

# utils
rich
tqdm
pyYAML
python-dotenv==1.0.0

# formats models
coremltools 
onnx 
onnx-simplifier 
onnxruntime
openvino-dev>=2022.3
# nvidia-tensorrt
# tensorflow 
# tensorflowjs